src/lambdas/axon-evidence-data-preparator/requirements.txt
boto3>=1.26.0
botocore>=1.29.0
requests>=2.28.0
urllib3>=1.26.0
zipstream-new>=1.1.8
